Possibility to send the full object to Marathon on update
Prior to this change setting the value of some keys of an application
(like 'health_checks' or 'constraints') from a collection that contains objects
to an empty collection and then calling `MarathonClient.update_app` to update the
application resulted in health checks or constraints never being updated.

The cause of this was that `MarathonClient.update_app` always called
`app.to_json(minmal=True)` which kicks out all empty collections.

This change lets the developer choose whether `minimal` should be `True` or `False`.
It retains the original behaviour through a default.